Some core cleanups
* [eAVSwitch] Fix missing close on error

* [eHdmiCEC] Add more logging and prepare for py3

* Fix compiler warnings

* Update enigma.cpp

+++ lib/gui/elistbox.cpp
@@ -71,7 +71,7 @@ void eListbox::allowNativeKeys(bool allow)
 		ePtr<eActionMap> ptr;
 		eActionMap::getInstance(ptr);
 		if (allow)
-			ptr->bindAction("ListboxActions", 0, 0, this);
+			ptr->bindAction("ListboxActions", (int64_t)0, 0, this);
 		else
 			ptr->unbindAction(this, 0);
 		m_native_keys_bound = allow;
